requirements.txt
torch==2.3.0
numpy==1.26.4
pandas==2.2.1
scikit-learn==1.4.2
yfinance==0.2.38
matplotlib==3.8.3
tqdm==4.66.1
pydantic==2.6.3
pyyaml==6.0.1
fastapi==0.110.0
uvicorn==0.29.0
